support sse chunked response
