+++ example/10-ai/sfn-currency-converter/main.go
@@ -80,7 +80,7 @@ func handler(ctx serverless.Context) {
 		// err = ctx.Write(0x61, append(reqID, []byte("can not understand the target currency")...))
 	} else {
 		// err = ctx.WriteWithTarget(0x61, []byte(fmt.Sprintf("The exchange rate of %s to USD is %f", msg.Target, rate)), "user-1")
-		result = fmt.Sprintf("The exchange rate of %s to USD is %f, compute result is %f", msg.Target, rate, msg.Amount*rate)
+		result = fmt.Sprintf("%f", msg.Amount*rate)
 		// err = ctx.Write(0x61, append(reqID, []byte(fmt.Sprintf("The exchange rate of %s to USD is %f, compute result is %f", msg.Target, rate, msg.Amount*rate))...))
 	}
