+++ pkg/bridge/ai/ai.go
@@ -135,6 +135,12 @@ func (a *AIServer) Serve() error {
 		// }
 
 		fmt.Fprintf(v.ResponseWriter, "data: %s\n\n", call.Arguments)
+		flusher, ok := v.ResponseWriter.(http.Flusher)
+		if ok {
+			flusher.Flush()
+		} else {
+			ylog.Debug("--------do not support flush-------------")
+		}
 	})
 
 	err = sfn.Connect()
@@ -233,6 +239,7 @@ func (a *AIServer) Serve() error {
 
 		// set Event Source response
 		w.Header().Set("Content-Type", "text/event-stream")
+		w.Header().Set("Transfer-Encoding", "chunked")
 
 		// wait for http response generated by all sfn-s
 		// or, timeout after 5 seconds
